refactor(peer_check): move verify_peer onto impl Allowlist; fold allowlist module into linux
- verify_peer is now `Allowlist::verify_peer(&self, stream)` instead of a
  free function with `(stream, allowlist)`. Reads more naturally at the
  call site (`self.allowlist.verify_peer(&stream)`) and groups the policy
  type with the operation that uses it.
- Fold the file-reading helpers back into peer_check/linux.rs and drop
  the separate allowlist submodule; the helpers are Linux-only anyway.
- Fix macos.rs `Alowlist::default()` typo that broke macOS clippy.
